+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+  * Entry points need no longer be syntactically first-order.
+
 ### Removed
 
 ### Changed